Update tmp.ino
This code works very well with the bluetooth RN-42. The temperature sensor data is send wirelessly to PC without changing anything. Check the ports before the upload of the sketch to microcontroller and also when the bluetooth is connected wirelessly. Always opt for the outgoing port during wireless connection. The receiving data is seen in the serial montor of arduino ide.